update router with path parameters
