+++ unchained/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod error;
 pub mod router;
+pub mod server;
 pub mod templates;
 pub mod workers;
